+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+v3.4.0
+----------
+* feat: add interactive and order message support in history models
+
 v3.3.0
 ----------
- * feat: add interactive and order message support in history models
+ * feat: implement history saving for completed streamed messages in gRPC
 
 v3.2.0
 ----------